test(DialogFragment): use empty fragment as dialog container
