+++ app/src/androidTest/java/com/github/ashutoshgngwr/noice/fragment/DialogFragmentTest.kt
@@ -1,8 +1,8 @@
 package com.github.ashutoshgngwr.noice.fragment
 
-import android.view.View
-import androidx.annotation.StringRes
-import androidx.test.core.app.ActivityScenario
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.testing.FragmentScenario
+import androidx.fragment.app.testing.launchFragmentInContainer
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.action.ViewActions.replaceText
@@ -11,16 +11,12 @@ import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.github.ashutoshgngwr.noice.EspressoX
-import com.github.ashutoshgngwr.noice.MainActivity
 import com.github.ashutoshgngwr.noice.R
 import com.github.ashutoshgngwr.noice.RetryTestRule
-import com.google.android.material.textfield.TextInputLayout
 import io.mockk.every
 import io.mockk.mockk
-import org.hamcrest.Description
 import org.hamcrest.Matchers.allOf
 import org.hamcrest.Matchers.not
-import org.hamcrest.TypeSafeMatcher
 import org.junit.Assert.assertEquals
 import org.junit.Before
 import org.junit.Rule
@@ -35,29 +31,19 @@ class DialogFragmentTest {
   @JvmField
   val retryTestRule = RetryTestRule(5)
 
-  private lateinit var activityScenario: ActivityScenario<MainActivity>
+  private lateinit var emptyFragmentScenario: FragmentScenario<Fragment>
   private lateinit var dialogFragment: DialogFragment
-
-  private fun hasErrorText(@Suppress("SameParameterValue") @StringRes expectedErrorText: Int) =
-    object : TypeSafeMatcher<View>() {
-      override fun describeTo(description: Description?) = Unit
-      override fun matchesSafely(item: View?): Boolean {
-        if (item !is TextInputLayout) return false
-        val error = item.error ?: return false
-        return item.context.getString(expectedErrorText) == error.toString()
-      }
-    }
 
   @Before
   fun setup() {
     dialogFragment = DialogFragment()
-    activityScenario = ActivityScenario.launch(MainActivity::class.java)
+    emptyFragmentScenario = launchFragmentInContainer(null, R.style.Theme_App)
   }
 
   @Test
   fun testTitleText() {
-    activityScenario.onActivity {
-      dialogFragment.show(it.supportFragmentManager) {
+    emptyFragmentScenario.onFragment {
+      dialogFragment.show(it.childFragmentManager) {
         title(android.R.string.yes)
       }
     }
@@ -68,8 +54,8 @@ class DialogFragmentTest {
 
   @Test
   fun testPositiveButton() {
-    activityScenario.onActivity {
-      dialogFragment.show(it.supportFragmentManager) {
+    emptyFragmentScenario.onFragment {
+      dialogFragment.show(it.childFragmentManager) {
         positiveButton(android.R.string.yes)
       }
     }
@@ -83,8 +69,8 @@ class DialogFragmentTest {
 
   @Test
   fun testNegativeButton() {
-    activityScenario.onActivity {
-      dialogFragment.show(it.supportFragmentManager) {
+    emptyFragmentScenario.onFragment {
+      dialogFragment.show(it.childFragmentManager) {
         negativeButton(android.R.string.no)
       }
     }
@@ -98,8 +84,8 @@ class DialogFragmentTest {
 
   @Test
   fun testMessageText() {
-    activityScenario.onActivity {
-      dialogFragment.show(it.supportFragmentManager) {
+    emptyFragmentScenario.onFragment {
+      dialogFragment.show(it.childFragmentManager) {
         message(android.R.string.yes)
       }
     }
@@ -113,8 +99,8 @@ class DialogFragmentTest {
     val mockValidator = mockk<(String) -> Int>()
     every { mockValidator.invoke("invalid") } returns android.R.string.no
     every { mockValidator.invoke("test") } returns 0
-    activityScenario.onActivity {
-      dialogFragment.show(it.supportFragmentManager) {
+    emptyFragmentScenario.onFragment {
+      dialogFragment.show(it.childFragmentManager) {
         input(
           hintRes = android.R.string.yes,
           preFillValue = "test",
@@ -128,7 +114,7 @@ class DialogFragmentTest {
       .perform(replaceText("invalid"))
 
     onView(withId(R.id.textInputLayout))
-      .check(matches(hasErrorText(android.R.string.no)))
+      .check(matches(EspressoX.withErrorText(android.R.string.no)))
 
     onView(withId(R.id.positive))
       .check(matches(not(isEnabled())))
@@ -137,7 +123,7 @@ class DialogFragmentTest {
       .perform(replaceText("test"))
 
     onView(withId(R.id.textInputLayout))
-      .check(matches(not(hasErrorText(android.R.string.no))))
+      .check(matches(not(EspressoX.withErrorText(android.R.string.no))))
 
     assertEquals("test", dialogFragment.getInputText())
   }
@@ -147,8 +133,8 @@ class DialogFragmentTest {
     var selectedItem = 0
     val items = arrayOf("test-0", "test-1", "test-2")
 
-    activityScenario.onActivity {
-      dialogFragment.show(it.supportFragmentManager) {
+    emptyFragmentScenario.onFragment {
+      dialogFragment.show(it.childFragmentManager) {
         singleChoiceItems(items, currentChoice = selectedItem) { choice ->
           selectedItem = choice
         }
